fix: in cypress tests change tos and privacy calls stubs, change exclude pattern for tests
