+++ packages/pn-pa-webapp/cypress/e2e/All.cy.ts
@@ -1,7 +1,7 @@
 import "./Auth.cy";
-// import "./TosAgreement.cy";
+import "./TosAgreement.cy";
 import "./FilterNotifications.cy";
-// import "./NewNotification.cy";
+import "./NewNotification.cy";
 import "./NewNotification.without.payment.methods.cy";
 import "./NotificationDetail.cy";
 import "./Pagination.cy";
