+++ packages/pn-pa-webapp/cypress/e2e/Pagination.cy.ts
@@ -7,9 +7,11 @@ describe("Pagination", () => {
   });
 
   beforeEach(() => {
+    // this prevents random errors in the app from breaking cypress tests
     Cypress.on('uncaught:exception', (err, runnable) => {
       return false;
     });
+
     cy.viewport(1920, 1080);
 
     // cy.intercept('GET', /delivery\/notifications\/sent/, {
@@ -18,14 +20,9 @@ describe("Pagination", () => {
       fixture: 'notifications/list-10/page-1'
     }).as('notifications');
 
-    cy.intercept(/TOS/, {
-      statusCode: 200,
-      fixture: 'tos/tos-accepted'
-    });
-    cy.intercept(/DATAPRIVACY/, {
-      statusCode: 200,
-      fixture: 'tos/privacy-accepted',
-    });
+    // stubs tos and privacy consents
+    cy.stubConsents();
+
     cy.visit('/dashboard');
   });
